Add back ManagedFile class
Partially revert "Remove sync command" (commit
`a66746602f09dd135bf2668ee89339a5b5bc0fe5`) in order to retain the
ManagedFile class and other code needed for this to work correctly.

Although this class was, and still is, unused in Underware, with
necessary changes for the initial version of Underware it will still be
needed; therefore re-adding it to live in Underware again.