Have configure and remove render genders directly
Rather than via Staging, which is a Metalware specific concept. After
doing this can also remove a whole load of now-unused Metalware-specific
code.

This commit does remove some of the code to handle the validation of the
rendered genders file before installation. All actual validation was
previously done at `sync` time, which is a Metalware concept and no
longer exists in Underware, but the specific code to validate a genders
file has been partially removed in this commit with the removal of the
Genders RenderMethod. At some point this should be added back, using
`NodeattrInterface.validate_genders_file` again, but as part of the new
process of rendering genders.