Remove Metalware design docs
These are specifically related to the design of Metalware and just kept
for future reference. Although some parts may still be relevant to
Underware other parts are also likely to be unrelated or misleading when
just considering Underware. Therefore removing these; if needed in
future we can always look in the git history or the Metalware repo to
see why decisions were made in Metalware.