Merge Underware Gemfile into Gem-generated files
When writing a Gem, dependencies should be specified in a `gemspec`
file; this can then be referenced from a `Gemfile` for development use.
This commit appropriately merges the current Underware Gemfile into
these files as needed; since Underware will be used both as a Gem and as
a stand-alone tool, we want slightly more code in the Gemfile to specify
the Ruby to use, dependency sources etc. than if it was just going to be
used as a Gem.