Lock down `activesupport` version used
Metalware currently has a dependency on Rails 5.1, and therefore
`activesupport` 5.1, and so adding Underware as a Metalware dependency
when Underware in turn depends on `activesupport` 5.2 was causing
dependency resolution to fail. Locking Underware to use `activesupport`
5.1 fixes this.

At some point it would be worth updating both Underware and Metalware to
use the newer `activesupport`, but don't see a real need to do this yet
and don't want to deal with updating Metalware's Rails version right now
(and risk spuriously breaking things).