Correct deployment URLs to be under `/metalware`
The previous `/underware` URLs didn't make sense and would never have a
chance of resolving, and were victims of earlier over-zealous `sed`-ing.
These URLs are all specifying places to retrieve files when doing a
Metalware build, and will be served by the Metalware deployment server
(if that's installed) and available under `/metalware`.