Use version number taken from `VERSION` constant
Rather than having this independently set in Cli class and by this
constant (Cli class' version is used by `underware --version`; constant
is picked up by Rubygems for Underware Gem). Also update release process
doc to account for updating this as part of a release.

The current version number for Underware has also been bumped to be the
correct number for the next Underware alpha release, and the format of
this has been changed to be a valid Gem version format (from `alpha-01`
to `0.2.0`).