Use glob to specify Underware Gem's files
Rather than depending on `git ls-files`. When Underware is installed in
a production environment, we currently just download and extract an
archive of the state of the repo as of the branch of Underware to be
installed, rather than cloning the Git repo itself (presumably this was
originally done to avoid Underware/Metalware requiring Git also be
installed, and/or as this is much quicker than cloning the repo with all
its history).

Therefore we cannot depend on Git being installed, but since we only
want to expose files in `lib/` we can glob this to expose all of these
files.