Test, and correct, Dependency enforcing errors
Add test coverage for some of the error messages which can be produced
by the Dependency class, and correct the repo dependency failure message
to account for Dependency now being able to be used by multiple tools.

Also document why there is a difference between the format of the
suggested commands to run in the error messages for `repo` vs
`configure` dependencies (see comments for full explanation).