Fix bugs with GroupCache accessor defaults
Previously the defaults for the various values to be read from the
GroupCache were being set at load time, but only if the current cache
file was empty/non-existent. If the current cache file had some content,
but not the key that is being retrieved, then the default would not be
set and confusing errors could occur at distant locations - this was
possible in our tests, where in at least one place we write a partial
group cache file, and would also occur if a user manually edited the
cache file in various ways.

This commit fixes these issues by ensuring the default is set within the
accessors themselves, if needed.