Have build_interface configured by question :tada:
Rather than having this come from the `server.yaml` file in the repo,
which needed to be manually edited in order to use a different
interface.

This is an improvement in any case, as it removes the need to manually
edit this file when building a cluster and means we can handle
configuring the build interface in the same way as any other
configurable value. And in particular we are doing this now as this
removes one more dependency of Underware on the repo, and so takes us a
step closer to removing the concept of an Underware repo entirely.