Tidy up HashMerger tests to be clearer
Rather than indirection where we use a function which yields to
different blocks containing case expression to assert what different
values of the merged namespaces are, let's just assert that the merged
data equals the overall hash that we expect it to be; I think this makes
things much more obvious as to what is actually going on.

Also improve various naming to not refer to something which isn't a hash
as a `merged_hash`, not indicate that the merged config is returned when
it is only part of the overall merged namespace etc.