Remove unneeded `Local` class and `Node#create`
There no longer seems to be a need to have a distinct `Local` class, as
this has no unique behaviour or data. Removing it makes things a bit
simpler; in particular this means we no longer need a confusing
`Node#create` method which should be used instead of `Node#new`, and
which functions identically apart from that it sometimes returns a
`Node` and sometimes a `Local` (which is still a `Node` and functionally
identical, just has a different class).