Add test coverage for namespace hash merging
There was not previously full, direct test coverage of how we were
passing through parameters to the underlying Config and Answer
HashMergers from the Domain/Group/Node HashMergerNamespaces, i.e.
essentially test coverage that the `hash_merger_input` method gives the
correct value (but without being coupled to this private method, which
is an implementation detail).

This commit adds direct coverage of this, to document how this works and
make it simpler to change. These tests should currently be pretty
straightforward but could probably be made less verbose.