Pass `platform` to root rather than individual namespaces
Rather than have the `platform` parameter be passed in on creation of
the individual Domain/Group/Node namespaces, instead have this passed
into the root Alces namespace and from there seamlessly used as needed
in the lower level namespaces.

This should allow us to configure an Alces namespace instance once and
then have all rendering and accessing of values against this or any part
of it automatically use the correct config merging, with or without the
platform config included, depending on whether this was initially
specified.