Have `template` only render for configured platforms
We should render all templates for a given platform `$platform` if and
only if the platform config file at `data/config/platforms/$platform`
exists (see
https://github.com/alces-software/underware/issues/19#issuecomment-444458544).

Now that we're rendering for each platform in turn, this also makes it
straightforward to only create a single Alces instance per platform,
which is better both to avoid duplication and for improved performance,
since all parts of an Alces instance for a platform will remain
constant, but can be expensive to recreate (e.g. repeatedly
loading/merging/rendering configs etc.).