Extract method for creating files within nested dirs
Because `template` works with arbitrarily nested templates, and renders
these to arbitrarily nested files in the rendered files directory, in
several places in both test and production code we were using `mkdir_p`
on the `dirname` for a given path and then touching or writing to the
file itself. Therefore extract a method for this simple but common
pattern, making this slightly simpler and clearer.