Never consider loopback for default build_interface
From
https://github.com/alces-software/underware/issues/25#issuecomment-443207771
point (1) we decided that it never makes sense to choose the loopback
interface when building a cluster. Therefore we now exclude this from
the list of interfaces that we select the default build interface from
when no build interface has been configured, and similarly this will not
be provided as an option for the `interface` question type once this has
been implemented.