Fail fast if no network interfaces available
If no network interfaces are available, we're not going to be able to
define a build interface for Underware and various things will break.

Since this would likely indicate something badly configured in the
user's environment (e.g. they've not got any networks configured), fail
fast if this ever occurs when they try to access the available
interface, so that they can attempt to fix this and we don't blow up
less descriptively elsewhere.