Simplify ad-hoc output handling in configurator tests
Rather than always redirecting stdout, and sometimes redirecting stderr
in different ways, we can instead always redirect both and just return
stderr from `configure_with_answers` (since this is the only one we want
to sometimes assert the presence of output in).

This simplifies various tests and means we should never get extraneous
output printed when the tests in this file are run, even if we don't
want to actually check this output in any given test.