Give consistent names to various FilePath methods
Give more consistent names to various file path methods - in general,
these do not need to end with `_file`, as by existing in this module it
is already clear that these are for file paths; also, methods for
directory paths (rather than for individual files) should indicate this
in their names by ending `_dir`.

One exception to the above is that the `namespace_data_file` method has
been left as-is, since these files are themselves considered 'namespace
data files'.