Simplify use of FileSystemConfigurator
There is no longer a need to have this passed in as an argument to
`FileSystem.test`, to specify which configurator is being used, or to
have a `reset` flag passed into a method to toggle its behaviour to
reset this, since there is now only ever a single
`FileSystemConfigurator` in use at a time.

Instead, we can just have the single root `FileSystemConfigurator`
always be used, and just add a method on to `FileSystem` to reset this,
thereby simplifying things.