Remove unnecessary indirection in Output
These methods were originally needed for the Metalware GUI, so we could
handle both displaying output in the GUI in different ways and
displaying to `stderr` in the CLI, depending on whether Metalware was
running in GUI or CLI mode. Now that the Metalware GUI is no more this
is just unnecessary indirection around outputting to `stderr`, and so
can be removed.