Thread through `eager_render` UnderwareRecursiveOpenStruct flag
This is threaded through from the `Alces` namespace at the top-level,
down to the various different individual `UnderwareRecursiveOpenStruct`
instances that we create. The only place a default value for the flag is
set is at the `Alces` level, which should hopefully ensure that this is
always passed through from this level downwards (and so the individual
`UnderwareRecursiveOpenStruct`s ultimately created should always
(hopefully) have the correct value set for the flag).

Currently we never actually create an `Alces` instance with this flag
passed, and so the default value will always be used, and so this commit
should not actually change any behaviour; however everything should now
be set up so this flag can be passed in to an `Alces` instance at the
top level and things should Just Work as intended.