Split executing the command out of initialize
The initialize method executes the command for legacy reason. This
isn't ideal as it makes it harder to reuse the command class. At this
point this can't be changed as to much relies on it

But instead, there is a `noop` flag that causes the initialize to skip
the start method entirely. This will allow an instance of the class to
be created without running it.

It can also be initialized without the args or options input as these
will need to go as well (TBA)