Use ClusterAttr to generate the nodes list
Previously `NodeattrInterface` was used to get the nodes array on
`alces`. It now uses the `ClusterAttr`

Some of the specs needed to be updated to create the nodes through
`ClusterAttr`. A couple of the older legacy specs still used a mocked
genders file. These specs are incompatible with the new mocking
mechanism and have been removed