Initialize `NodePrototype` with optional genders
The `NodePrototype` is an abstraction to nodes that do not exist yet. It
will be used to generate answer hashes during the `configure` stage.

As the nodes do not exist yet, it needs to be initialized with a genders
input. This needs to be optional b/c the `Node` namespace does not use
it.

The `Node` input redefines the `genders` method as this delays the
initialization of the `genders` array until it is called