Patch `TTY::Config` to allow `YAML` aliases
The bug this fixes is rather esoteric. No library is particular at
fault, but its due to how everything fits together

It all start with the `ClusterAttr.add_nodes` method. When it adds
multiple nodes in one method call, it assigns `groups` which are saved
in `YAML`.

Now the underlining `Psych` library kicks in and tries to be smart. B/c
its saving the same groups array object, it uses `YAML` anchors:

```
nodes:
  local: []
  node01: &1
  - group1
  - login
  - my-cluster
  - all
  node02: *1
  node03: *1
  node04: *1
```

Now this is where `TTY::Config` tries to be smart and `safe_load` the
`YAML`. Now by default the `safe_load` method does not allow aliases
unless explicitly told to.

This causes a problem b/c the `YAML.safe_load` method is buried in a big
long method call. Whilst it would be possible to override this method,
it would be very brittle if it ever changed.

Also it isn't a great idea patching `YAML` globally as this could break
other libraries. Instead the `TTY::Config::YAML` constant is redefined
to be the patch. This keeps it local to within `TTY::Config`

Now the patch itself is rather simple, it redefines `safe_load` to
always allow aliases