Remove the `local` node from the configurator
The local node is no longer a concept and should not be configurable.
It has been removed from the `configure.yaml`.

A few of the specs needed updating as they assumed there would be a
local section that needed configuring. This was mostly a case of
removing the local specific code