Cache the cluster paths on `FilePath`
This is to manage the migration from static `FilePath` to dynamic
`DataPath`s. To handle this, there is a currently in scope cluster
`Commands::Init::CLUSTER_IDENTIFIER`

Eventually cluster switching functionality will be added to change this.
However before this can be implemented, the `FilePath` methods need to
start using the dynamic methods