Allow DataPath to have an arbitrary base path
The Plugin directory mimics the structure of `data` directory. As such
it makes sense to handle them in a similar manner. The spec did need
updating to match the newer `configs` layout with the `nodes` and
`groups` sub-directories.

This will break the existing plugins. However plugins are going to be
revamped at somepoint anyway.