Port the generation of the `answer` paths
These paths are generated dynamically as they follow a specific pattern.
There is always a `domain`, `group`, and `node` method. An additional
`platform` method has been added as it fits the pattern (even if there
is no answers for platforms).

This will allow the pattern to be used for `configs`