Remove some of the caching of the `DataPath`
The `alces` object contains a bunch of paths and shouldn't be cached
between commands and thus this has been removed.

Also the `FilePath` class can not cache the `DataPath` as this prevents
the paths from changing. Instead each call to `FilePath` needs to create
a new `DataPath` class. This allows the paths to change when the cluster
is switched