Deprecate the usage of `DataCopy.init_cluster`
This is a bit of a weird method as it creates and instance and then
calls `all` on it. By calling it `init_cluster` it also had to handle
`layouts`, thus adding complexity.

Instead `layout_to_cluster` should be used instead. It doesn't have any
pretext that it creates a full cluster