Skip the royally broken specs that are hard to fix
A bunch of the specs are broken b/c they do not configure a cluster
before running. A bunch of these specs are for features that should be
removed, so there isn't much point fixing them ATM.

This mostly pertains to `command` specs as they run the check. They will
need to be reviewed at some point