Revamp the dependency mechanism between commands
It isn't uncommon to require a certain condition to be met before a
command can be ran. Previously this use the convoluted
`DependencySpecifications` to enforce them.

Instead a simple list of `dependencies` are maintained as a class
instance variable. As a test case, the cluster assertion is implemented
in this manner