Include the mixin instead of extending
This makes a difference to the ruby metaproggramming. Because `double`
returns an instance, it can not be directly extended.

Instead an anonymous double class is created instead. This allows it to
be `included` before stubbing the methods.
NOTE: The standard `rspec` doubles can't be used as they do not actually
create the methods. Instead they use `method_missing`