Force the removal of the answer files
Ruby handles `rm` in a funny manner as it never asks for confirmation.
So in practice it is always a force deletion.

However `rm` will still fail if the file does not exist. This is why
`rm_f` needs to be used instead. It doesn't fail if the file doesn't
exist