Turn the helper render paths into a single method
Working out where to store a rendered file is a bit complicated as it
depends on the platform and if it is a core file.

There are also differences between the node/group/domain sections. For
simplicity, there is only a single method that handles all the use
cases. The inputs are set using flags