Remove the gemspec file from architect
It is unusual to commit both the lock and gemspec file. This causes all
sorts of issues with architects version number as it gets hard coded
into the lock file. This causes the lock file to become stale and allows
version bugs to silently creep in

As architect is no longer being used as a gem, the gemspec file can be
removed completely. This means architect/underware version number no
longer appears in the lock file, thus preventing it from becoming stale